get user withdrawal/deposit in wallet
