+++ tools/user.js
@@ -1042,95 +1042,6 @@ const getUserAudits = (userId, limit, page, orderBy, order, startDate, endDate, 
 		});
 };
 
-const getTransactions = (
-	type,
-	kitId,
-	currency,
-	status,
-	dismissed,
-	rejected,
-	processing,
-	waiting,
-	limit,
-	page,
-	orderBy,
-	order,
-	startDate,
-	endDate,
-	format
-) => {
-	let promiseQuery;
-	if (kitId) {
-		if (type === 'deposit') {
-			promiseQuery = getUserByKitId(kitId, false)
-				.then((user) => {
-					return getNodeLib().getDeposits(user.network_id, currency, status, dismissed, rejected, processing, waiting, limit, page, orderBy, order, startDate, endDate);
-				});
-		} else if (type === 'withdrawal') {
-			promiseQuery = getUserByKitId(kitId, false)
-				.then((user) => {
-					return getNodeLib().getWithdrawals(user.network_id, currency, status, dismissed, rejected, processing, waiting, limit, page, orderBy, order, startDate, endDate);
-				});
-		}
-	} else {
-		if (type === 'deposit') {
-			promiseQuery = getNodeLib().getDeposits(undefined, currency, status, dismissed, rejected, processing, waiting, limit, page, orderBy, order, startDate, endDate);
-		} else if (type === 'withdrawal') {
-			promiseQuery = getNodeLib().getWithdrawals(undefined, currency, status, dismissed, rejected, processing, waiting, limit, page, orderBy, order, startDate, endDate);
-		}
-	}
-	return promiseQuery
-		.then((transactions) => {
-			if (format) {
-				if (transactions.data.length === 0) {
-					throw new Error(NO_DATA_FOR_CSV);
-				}
-				const csv = parse(transactions.data, Object.keys(transactions.data[0]));
-				return csv;
-			} else {
-				return transactions;
-			}
-		});
-};
-
-const getUserDepositsByKitId = (
-	kitId,
-	currency,
-	status,
-	dismissed,
-	rejected,
-	processing,
-	waiting,
-	limit,
-	page,
-	orderBy,
-	order,
-	startDate,
-	endDate,
-	format
-) => {
-	return getTransactions('deposit', kitId, currency, status, dismissed, rejected, processing, waiting, limit, page, orderBy, order, startDate, endDate, format);
-};
-
-const getUserWithdrawalsByKitId = (
-	kitId,
-	currency,
-	status,
-	dismissed,
-	rejected,
-	processing,
-	waiting,
-	limit,
-	page,
-	orderBy,
-	order,
-	startDate,
-	endDate,
-	format
-) => {
-	return getTransactions('withdrawal', kitId, currency, status, dismissed, rejected, processing, waiting, limit, page, orderBy, order, startDate, endDate, format);
-};
-
 const checkUsernameIsTaken = (username) => {
 	return getModel('user').count({ where: { username }})
 		.then((count) => {
@@ -1314,8 +1225,6 @@ module.exports = {
 	toggleFlaggedUserById,
 	getUserLogins,
 	getUserAudits,
-	getUserDepositsByKitId,
-	getUserWithdrawalsByKitId,
 	setUsernameById,
 	getAffiliationCount,
 	isValidUsername,
